+++ artiq/frontend/aqctl_lda.py
@@ -23,7 +23,7 @@ def get_argparser():
                              "available device will be used.")
     parser.add_argument("--simulation", action="store_true",
                         help="Put the driver in simulation mode.")
-    verbosity_args(parser)
+    add_common_args(parser)
     return parser