+++ artiq/frontend/artiq_client.py
@@ -21,7 +21,7 @@ from artiq.protocols.pc_rpc import Client
 from artiq.protocols.sync_struct import Subscriber
 from artiq.protocols.broadcast import Receiver
 from artiq.protocols import pyon
-from artiq.tools import short_format, parse_arguments
+from artiq.tools import short_format, add_common_args, parse_arguments
 
 
 def clear_screen():
@@ -64,10 +64,7 @@ def get_argparser():
                                  "(defaults to head, ignored without -R)")
     parser_add.add_argument("-c", "--class-name", default=None,
                             help="name of the class to run")
-    parser_add.add_argument("-v", "--verbose", default=0, action="count",
-                            help="increase logging level of the experiment")
-    parser_add.add_argument("-q", "--quiet", default=0, action="count",
-                            help="decrease logging level of the experiment")
+    add_common_args(parser)
     parser_add.add_argument("file", metavar="FILE",
                             help="file containing the experiment to run")
     parser_add.add_argument("arguments", metavar="ARGUMENTS", nargs="*",